requirements.txt
confluent-kafka
faker
python-dotenv
boto3
